+++ alfalfa/optimizer/__init__.py
@@ -1,2 +1,3 @@
 from .gbm_model import GbmModel
 from .proposals import propose
+from .opt_model import build_opt_model
