+++ alfalfa/utils/plots/__init__.py
@@ -1,0 +1,3 @@
+from .plot_gp import plot_gp_1d, plot_gp_2d, plot_covar_matrix
+from .plot_training import plot_loss_logs
+from .plot_trees import plot_forest, plot_tree
