+++ alfalfa/utils/plots/plot_gp.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 from typing import Callable
 
-from .logger import Logger
+from ..logger import Logger
 
 def plot_gp_1d(model: gpy.models.ExactGP, test_x, target: Callable, ax=None):
     if ax is None:
@@ -77,22 +77,3 @@ def plot_covar_matrix(model: gpy.models.ExactGP, test_x, ax=None):
     ax.set_xticks([])
     ax.set_yticks([])
     return ax, im
-
-
-def plot_loss_logs(logger: Logger, loss_key: str, step_key: str, test_loss_key: str):
-    loss = np.array(logger[loss_key])
-    steps = np.array(logger[step_key])
-    fig, ax = plt.subplots()
-    for step_type in np.unique(steps):
-        filtered_loss = np.where(steps == step_type, loss, np.nan)
-        ax.plot(filtered_loss, label=step_type)
-
-    test_loss_xs = np.concatenate((
-        np.array([0]),
-        (1 + np.argwhere(steps[1:] != steps[:-1]).flatten()),
-        np.array([steps.shape[0]]),
-    ))
-    ax.plot(test_loss_xs, logger[test_loss_key], label="test loss")
-    ax.legend()
-
-    return fig, ax
