+++ examples/bayes_opt/run_lgbm_study.py
@@ -5,13 +5,13 @@ import lightgbm as lgb
 from botorch.utils import standardize
 import matplotlib.pyplot as plt
 
-from alfalfa.leaf_gp.bb_func_utils import get_func
+from alfalfa.utils.bb_funcs import get_func
 from alfalfa import AlfalfaForest
 from alfalfa.tree_models.tree_kernels import AlfalfaGP
 from alfalfa.optimizer import propose, build_opt_model
-from alfalfa.leaf_gp.gbm_model import GbmModel
+from alfalfa.optimizer.gbm_model import GbmModel
 from alfalfa.tree_models.lgbm_tree import lgbm_to_alfalfa_forest, fit_leaf_gp
-from alfalfa.leaf_gp.space import Space
+from alfalfa.utils.space import Space
 from alfalfa.utils.plots import plot_gp_1d
 
 from argparse import ArgumentParser
